*sRGB render buffer complite *Fix max (and default) gamma scale value
git-svn-id: https://svn.code.sf.net/p/quake2xp/code@1004 d83f5309-02f3-4a42-a5f4-12e6af65da40

+++ trunk/ref_gl/r_light.c
@@ -1938,7 +1938,7 @@ qboolean R_MarkLightLeaves (worldShadowLight_t *light) {
 	light->area = CM_LeafArea (leafnum);
 
 	if (!light->area) {
-	skip:	Com_DPrintf ("Out of BSP, rejected light at %f %f %f\n", light->origin[0], light->origin[1], light->origin[2]);
+	skip:	Com_DPrintf ("Out of BSP, rejected light at %i %i %i\n", (int)light->origin[0], (int)light->origin[1], (int)light->origin[2]);
 		return qfalse;
 	}
